refactoring in progress

+++ src/main/java/ru/javarush/cryptoanalyser/parfenov/commands/AbstractCrypt.java
@@ -32,14 +32,9 @@ public abstract class AbstractCrypt {
         return new Result(ResultCode.OK, "Done");
     }
     public int encrypting(int character, int key, String alphabet) {
-        //String alphabet = Alphabet.TINY_ALPHA;
         int a = alphabet.indexOf((char)character);
         if(a == -1) return a;
         int b = Math.floorMod((a + (key % alphabet.length())), alphabet.length());
-        return (int) alphabet.charAt(b);
+        return alphabet.charAt(b);
     }
-
-
-
-
 }
